Data sizes, job duration and reporting layout.
Moved $session.BackupStats.BackupSize and $session.BackupStats.DataSize to $JobSize and $TransfSize for ease of use in calculations.
Series' of if statements then use these to determine the best unit in which to report the job and transferred data sizes. If there's no match (unlikely) then it fallback to reporting in bytes.

Added job duration to the report.

Changed emojis from an angry face for warning, green heart for success and fire for failure. This appears rather unprofessional and the meaning is not immediately obvious. Warning is now a yellow warning triangle, success is a green check-mark, and failure is a red cross.

Changed report section titles (Job, Status, Details) from italic to bold. It didn't make much sense for a title to be italic and having it bold makes the report much more clear.